refactor: move master-detail endpoint to dedicated controller
